+++ src/ast/cdl/actions_and_functions/mod.rs
@@ -2,10 +2,10 @@ mod action_declaration_term;
 mod action_term;
 mod function_term;
 mod param_term;
-mod returns_term;
+mod returns_declaration_term;
 
 pub use action_declaration_term::*;
 pub use action_term::*;
 pub use function_term::*;
 pub use param_term::*;
-pub use returns_term::*;
+pub use returns_declaration_term::*;
